move lib files to subfolder
