+++ packages/deploy-tools/src/index.ts
@@ -16,6 +16,6 @@
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
 
-export * from './deploy-utils';
-export * from './hockey';
-export * from './s3';
+export * from './lib/deploy-utils';
+export * from './lib/hockey';
+export * from './lib/s3';
